blog_backend/requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.3
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
psycopg2==2.9.3
Pygments==2.12.0
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.1.0
